Test inverted order rolling/quantiles
Helper functions just added to gSPEI take a quantile of a full time series, and then plot shows the rolling mean of that quantile.  Now added tester code to first take rolling means per model (which is what our previously-submitted figures plot) and then compute a quantile of those rolling means.  The result is more intuitive to interpret in overlay.