Create highlight_severity_fig.py
Add script to plot the figure uploaded as "Key Figure" for ESD: a single panel of glacial buffering on drought severity at end of century.  Compared with Figure 3 of manuscript, this figure focuses on a single time slice, has simpler axes labels, and shades the positive-buffering region of the plot.