Revise basin read-in for CMIP6 SPEI
Basins were previously read in same order as Finn's work / CMIP5 approach.  Sloan's output for CMIP6 had slightly different order and spelling.  After we discovered this in the precip analysis, checked the CMIP6 SPEI read-in imagining it might have the same problem.  With the revised read-in, expected relationships (to percent glacier cover and to aridity index) return to the plots.  Move forward with this.